# Conflicts:
#	obs-studio-server/source/nodeobs_api.cpp
#	obs-studio-server/source/nodeobs_settings.cpp

+++ obs-studio-client/source/scene.cpp
@@ -353,11 +353,11 @@ Nan::NAN_METHOD_RETURN_TYPE osn::Scene::GetItems(Nan::NAN_METHOD_ARGS_TYPE info)
 	if (!ValidateResponse(response))
 		return;
 
-	auto arr = Nan::New<v8::Array>(response.size() - 1);
+	auto arr = Nan::New<v8::Array>(int(response.size()) - 1);
 
 	for (size_t i = 1; i < response.size(); i++) {
 		osn::SceneItem* obj = new osn::SceneItem(response[i].value_union.ui64);
-		Nan::Set(arr, i - 1, osn::SceneItem::Store(obj));
+		Nan::Set(arr, uint32_t(i - 1), osn::SceneItem::Store(obj));
 	}
 
 	info.GetReturnValue().Set(arr);
@@ -387,11 +387,11 @@ Nan::NAN_METHOD_RETURN_TYPE osn::Scene::GetItemsInRange(Nan::NAN_METHOD_ARGS_TYP
 	if (!ValidateResponse(response))
 		return;
 
-	auto arr = Nan::New<v8::Array>(response.size() - 1);
+	auto arr = Nan::New<v8::Array>(int(response.size() - 1));
 
 	for (size_t i = 1; i < response.size(); i++) {
 		osn::SceneItem* obj = new osn::SceneItem(response[i].value_union.ui64);
-		Nan::Set(arr, i - 1, osn::SceneItem::Store(obj));
+		Nan::Set(arr, uint32_t(i - 1), osn::SceneItem::Store(obj));
 	}
 
 	info.GetReturnValue().Set(arr);
